pyproject.toml
[tool.poetry]
name = "covid19"
version = "0.1.1"
description = "Covid19 Data Analysis"
authors = ["Eric John <me@ericjohn.in>"]

[tool.poetry.dependencies]
python = "^3.6.1"
matplotlib = "^3.2"
aiohttp = "^3.6"
pandas = "^1.0"
flask = "^1.1"
python-dotenv = "^0.12.0"


[tool.poetry.dev-dependencies]
pytest = "^5.4"
pylint = "^2.4"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
